+++ spec/support/sandboxing.rb
@@ -2,8 +2,13 @@ require 'circleci/coverage_reporter/sandbox'
 
 RSpec.configure do |c|
   c.around do |example|
+    original_env = ENV.to_h
     CircleCI::CoverageReporter::Sandbox.sandboxed do |_config|
-      example.run
+      Dir.mktmpdir do |dir|
+        ENV['CIRCLE_ARTIFACTS'] = dir
+        example.run
+      end
     end
+    ENV.replace(original_env)
   end
 end